Factor out schemaRequestService
